Refactored `src/ui/loading-screen.ts` into multiple files to improve code organization and maintainability.
- Extracted types and constants to `loading-screen-types.ts`
- Moved the `LoadingScreen` UI class implementation to `loading-screen-ui.ts`
- Moved global APIs and legacy hooks to `loading-screen-progress.ts`
- Re-exported all modules via a barrel file at `loading-screen.ts`
- Updated `plan.md` and `REFACTORING_PLAN_REMAINING.md` status
