+++ REFACTORING_PLAN_REMAINING.md
@@ -2,15 +2,15 @@
 
 This document outlines the remaining 5 files (5,275 lines) that need to be refactored to break them into smaller, more manageable modules. This is a continuation of the large file refactoring task.
 
-**Status**: 50% Complete (5 of 10 files refactored)
+**Status**: 60% Complete (6 of 10 files refactored)
 
 ## Remaining Files Summary
 
 | File | Lines | Target Split | Status |
 |------|-------|--------------|--------|
 | src/world/generation.ts | 1078 | 4 files | ✅ DONE |
-| src/systems/region-manager.ts | 1064 | 3 files | ⏳ TODO |
-| src/ui/loading-screen.ts | 1063 | 3 files | ⏳ TODO |
+| src/systems/region-manager.ts | 1064 | 3 files | ✅ DONE |
+| src/ui/loading-screen.ts | 1063 | 3 files | ✅ DONE |
 | src/audio/audio-system.ts | 1047 | 3 files | ⏳ TODO |
 | src/ui/analytics-debug.ts | 1023 | 3 files | ⏳ TODO |
 | **TOTAL** | **5,275** | **16 files** | **⏳ 0%** |
